+++ engine/src/execution_graph/logic_controllers/BatchWindowFunctionProcessing.h
@@ -13,8 +13,9 @@ using ral::cache::kernel;
 using ral::cache::kernel_type;
 
 /**
- * @brief This kernel computes the main Window Function (ROW_NUMBER, LAG, LEAD, MIN, ...) to each batch already pattitioned
- * A new column should be added to the batch
+ * @brief This kernel computes the main Window Function (ROW_NUMBER, LAG, LEAD, MIN, ...)
+ * to each batch already pattitioned and sorted
+ * New columns will be added to each batch
  */
 
 class ComputeWindowKernel : public kernel {
@@ -23,7 +24,10 @@ public:
 		std::shared_ptr<Context> context,
 		std::shared_ptr<ral::cache::graph> query_graph);
 
-	std::unique_ptr<CudfColumn> compute_column_from_window_function(cudf::table_view input_cudf_view, cudf::column_view input_col_view, std::size_t pos);
+	std::unique_ptr<CudfColumn> compute_column_from_window_function(
+		cudf::table_view input_cudf_view,
+		cudf::column_view input_col_view,
+		std::size_t pos, int & agg_param_count);
 
 	std::string kernel_name() { return "ComputeWindow";}
 
@@ -34,9 +38,15 @@ public:
 	kstatus run() override;
 
 private:
-	std::vector<int> column_indices_partitioned;   // column indices to be partitioned
-	std::vector<int> column_indices_wind_func;     // column indices to be agg
-	std::vector<AggregateKind> aggs_wind_func; 
+	// LogicalComputeWindow(min_keys=[MIN($0) OVER (PARTITION BY $1)], lag_col=[LAG($0, 5) OVER (PARTITION BY $1)], n_name=[$2])
+	std::vector<int> column_indices_partitioned;   // column indices to be partitioned: [1]
+	std::vector<int> column_indices_to_agg;        // column indices to be agg: [0, 0]
+	std::vector<int> agg_param_values;     		   // due to LAG or LEAD: [5]
+	int preceding_value;     	   // X PRECEDING
+	int following_value;     		   // Y FOLLOWING
+	std::string frame_type;                        // ROWS or RANGE
+	std::vector<std::string> type_aggs_as_str;     // ["MIN", "LAG"]
+	std::vector<AggregateKind> aggs_wind_func;     // [AggregateKind::MIN, AggregateKind::LAG]
 };
 
 } // namespace batch