+++ engine/src/execution_graph/logic_controllers/LogicalProject.cpp
@@ -17,6 +17,7 @@
 #include "LogicalProject.h"
 #include "utilities/transform.hpp"
 #include "Interpreter/interpreter_cpp.h"
+#include "parser/expression_utils.hpp"
 
 namespace ral {
 namespace processor {
@@ -774,7 +775,11 @@ std::vector<std::unique_ptr<ral::frame::BlazingColumn>> evaluate_expressions(
     const std::vector<std::string> & expressions) {
     using interops::column_index_type;
 
-    std::vector<std::unique_ptr<ral::frame::BlazingColumn>> out_columns(expressions.size());
+    // Let's clean all the expressions that contains Window functions (if exists)
+    // as they should be updated with new indices
+    std::vector<std::string> new_expressions = clean_window_function_expressions(expressions, table.num_columns());
+
+    std::vector<std::unique_ptr<ral::frame::BlazingColumn>> out_columns(new_expressions.size());
 
     std::vector<bool> column_used(table.num_columns(), false);
     std::vector<std::pair<int, int>> out_idx_computed_idx_pair;
@@ -783,8 +788,8 @@ std::vector<std::unique_ptr<ral::frame::BlazingColumn>> evaluate_expressions(
     std::vector<cudf::mutable_column_view> interpreter_out_column_views;
 
     function_evaluator_transformer evaluator{table};
-    for(size_t i = 0; i < expressions.size(); i++){
-        std::string expression = replace_calcite_regex(expressions[i]);
+    for(size_t i = 0; i < new_expressions.size(); i++){
+        std::string expression = replace_calcite_regex(new_expressions[i]);
         expression = expand_if_logical_op(expression);
 
         parser::parse_tree tree;
@@ -887,9 +892,9 @@ std::vector<std::unique_ptr<ral::frame::BlazingColumn>> evaluate_expressions(
         out_columns.clear();
         computed_columns.clear();
 
-        size_t const half_size = expressions.size() / 2;
-        std::vector<std::string> split_lo(expressions.begin(), expressions.begin() + half_size);
-        std::vector<std::string> split_hi(expressions.begin() + half_size, expressions.end());
+        size_t const half_size = new_expressions.size() / 2;
+        std::vector<std::string> split_lo(new_expressions.begin(), new_expressions.begin() + half_size);
+        std::vector<std::string> split_hi(new_expressions.begin() + half_size, new_expressions.end());
         auto out_cols_lo = evaluate_expressions(table, split_lo);
         auto out_cols_hi = evaluate_expressions(table, split_hi);
 
@@ -921,10 +926,7 @@ std::unique_ptr<ral::frame::BlazingTable> process_project(
   const std::string & query_part,
   blazingdb::manager::Context * /*context*/) {
 
-    std::string combined_expression = query_part.substr(
-        query_part.find("(") + 1,
-        (query_part.rfind(")") - query_part.find("(")) - 1
-    );
+    std::string combined_expression = get_query_part(query_part);
 
     std::vector<std::string> named_expressions = get_expressions_from_expression_list(combined_expression);
     std::vector<std::string> expressions(named_expressions.size());